requirement.txt
python-dotenv==0.21.1
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.0
bcrypt==4.0.1
